reafctor: move shared to components
