+++ src/components/shared/errorBoundary/ErrorBoundary.module.scss
