+++ src/components/shared/loader/loader.module.scss
