+++ src/components/shared/logo/assets/logo.png
