+++ src/pages/main/components/cards-block/cards-block.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router';
 import { CardsList } from '../../../../components/cards-list/cards-list';
-import ErrorBoundary from '../../../../shared/errorBoundary/ErrorBoundary';
+import ErrorBoundary from '../../../../components/shared/errorBoundary/ErrorBoundary';
 import { SpacecraftsResponse } from '../../../../api/types';
 
 import styles from './cards-block.module.scss';
