+++ src/pages/main/components/error-button/error-button.tsx
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import { Button } from '../../../../shared/button/button';
+import { Button } from '../../../../components/shared/button/button';
 
 type Props = object;
